+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,13 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  For ESP32
+
+  AsyncTCP_SSL is a library for the ESP32
+
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncTCP_SSL
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"