+++ examples/multiFileProject/multiFileProject.h
@@ -1,0 +1,16 @@
+/****************************************************************************************************************************
+  multiFileProject.h
+  For ESP32
+
+  AsyncTCP_SSL is a library for the ESP32
+
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncTCP_SSL
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#pragma once
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "AsyncTCP_SSL.hpp"